refactor components and clean up
